requirements.txt
yfinance~=0.2.4
python-dotenv~=0.21.1